[DOES NOT COMPILE] Updating runtime console
Working on the runtime console window. Played around with NConsole. It almost does what I want but not quite. This wouldn't be the first time I've tried to work with command line parsing, so I've decided to make my own parsing library based off of NConsole. There are still some relics of NConsole in the application that will be removed next commit.